// src/API.ts
export type ModelStringKeyConditionInput = {
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
};

export type ModelRACSensorsFilterInput = {
  CO2?: ModelFloatInput | null,
  MAC?: ModelIntInput | null,
  O2?: ModelFloatInput | null,
  atm_temperature?: ModelFloatInput | null,
  conductivity?: ModelFloatInput | null,
  humidity?: ModelFloatInput | null,
  pH?: ModelFloatInput | null,
  temperature?: ModelFloatInput | null,
  timestamp?: ModelStringInput | null,
  and?: Array< ModelRACSensorsFilterInput | null > | null,
  or?: Array< ModelRACSensorsFilterInput | null > | null,
  not?: ModelRACSensorsFilterInput | null,
};

export type ModelFloatInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export enum ModelAttributeTypes {
  binary = "binary",
  binarySet = "binarySet",
  bool = "bool",
  list = "list",
  map = "map",
  number = "number",
  numberSet = "numberSet",
  string = "string",
  stringSet = "stringSet",
  _null = "_null",
}


export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelSizeInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelRACSensorsConnection = {
  __typename: "ModelRACSensorsConnection",
  items:  Array<RACSensors | null >,
  nextToken?: string | null,
};

export type RACSensors = {
  __typename: "RACSensors",
  CO2?: number | null,
  MAC: number,
  O2?: number | null,
  atm_temperature?: number | null,
  conductivity?: number | null,
  humidity?: number | null,
  pH?: number | null,
  temperature?: number | null,
  timestamp: string,
};
// ...
export type SoilSensors = {
  __typename: "SoilSensors",
  CO2?: number | null,
  MAC: number,
  O2?: number | null,
  atm_temperature?: number | null,
  conductivity?: number | null,
  humidity?: number | null,
  Nitrogen?: number | null,
  Phosphorous?: number | null,
  pH?: number | null,
  temperature?: number | null,
  timestamp: string,
};
// ...
export type ModelSoilSensorsFilterInput = {
  CO2?: ModelFloatInput | null,
  MAC?: ModelIntInput | null,
  O2?: ModelFloatInput | null,
  atm_temperature?: ModelFloatInput | null,
  conductivity?: ModelFloatInput | null,
  humidity?: ModelFloatInput | null,
  Nitrogen?: ModelFloatInput | null,
  Phosphorous?: ModelFloatInput | null,
  pH?: ModelFloatInput | null,
  temperature?: ModelFloatInput | null,
  timestamp?: ModelStringInput | null,
  and?: Array< ModelSoilSensorsFilterInput | null > | null,
  or?: Array< ModelSoilSensorsFilterInput | null > | null,
  not?: ModelSoilSensorsFilterInput | null,
};

export type ModelSoilSensorsConnection = {
  __typename: "ModelSoilSensorsConnection",
  items:  Array<SoilSensors | null >,
  nextToken?: string | null,
};
// ...
export type ListRACSensorsQueryVariables = {
  MAC?: number | null,
  timestamp?: ModelStringKeyConditionInput | null,
  filter?: ModelRACSensorsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListRACSensorsQuery = {
  listRACSensors?:  {
    __typename: "ModelRACSensorsConnection",
    items:  Array< {
      __typename: "RACSensors",
      CO2?: number | null,
      MAC: number,
      O2?: number | null,
      atm_temperature?: number | null,
      conductivity?: number | null,
      humidity?: number | null,
      pH?: number | null,
      temperature?: number | null,
      timestamp: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type ListSoilSensorsQueryVariables = {
  MAC?: number | null,
  timestamp?: ModelStringKeyConditionInput | null,
  filter?: ModelSoilSensorsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListSoilSensorsQuery = {
  listSoilSensors?:  {
    __typename: "ModelSoilSensorsConnection",
    items:  Array< {
      __typename: "SoilSensors",
      CO2?: number | null,
      MAC: number,
      O2?: number | null,
      atm_temperature?: number | null,
      conductivity?: number | null,
      humidity?: number | null,
      Nitrogen?: number | null,
      Phosphorous?: number | null,
      pH?: number | null,
      temperature?: number | null,
      timestamp: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};